+++ src/helpers/Db.php
@@ -933,6 +933,121 @@ class Db
     }
 
     /**
+     * Creates and executes a SQL statement for dropping an index if it exists.
+     *
+     * @param string $table The table that the index was created for.
+     * @param string|string[] $columns The column(s) that are included in the index. If there are multiple columns, separate them
+     * by commas or use an array.
+     * @param bool $unique Whether the index has a UNIQUE constraint.
+     * @param Connection|null $db The database connection.
+     * @since 4.0.0
+     */
+    public static function dropIndexIfExists(string $table, $columns, bool $unique = false, ?Connection $db = null): void
+    {
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        $indexName = static::findIndex($table, $columns, $unique, $db);
+
+        if ($indexName) {
+            $db->createCommand()
+                ->dropIndex($indexName, $table)
+                ->execute();
+        }
+    }
+
+    /**
+     * Creates and executes a SQL statement for dropping a foreign key if it exists.
+     *
+     * @param string $table The table that the foreign key was created for.
+     * @param string|string[] $columns The column(s) that are included in the foreign key. If there are multiple columns, separate them
+     * by commas or use an array.
+     * @param Connection|null $db The database connection.
+     * @since 4.0.0
+     */
+    public static function dropForeignKeyIfExists(string $table, $columns, ?Connection $db = null): void
+    {
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        $fkName = static::findForeignKey($table, $columns, $db);
+
+        if ($fkName) {
+            $db->createCommand()
+                ->dropForeignKey($fkName, $table)
+                ->execute();
+        }
+    }
+
+    /**
+     * Drops all the foreign keys that reference a table.
+     *
+     * @param string $table The table that the foreign keys should reference.
+     * @param Connection|null $db The database connection.
+     * @since 4.0.0
+     */
+    public static function dropAllForeignKeysToTable(string $table, ?Connection $db = null): void
+    {
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        $schema = $db->getSchema();
+        $rawTableName = $schema->getRawTableName($table);
+
+        foreach ($schema->getTableSchemas() as $otherTable) {
+            foreach ($otherTable->foreignKeys as $fkName => $fk) {
+                if ($fk[0] === $rawTableName) {
+                    $db->createCommand()
+                        ->dropForeignKey($fkName, $otherTable->name)
+                        ->execute();
+                }
+            }
+        }
+    }
+
+    /**
+     * Renames a table and its corresponding sequence (if PostgreSQL).
+     *
+     * @param string $table The table to be renamed.
+     * @param string $newName The new table name.
+     * @param Connection|null $db The database connection.
+     * @since 4.0.0
+     */
+    public static function renameTable(string $table, string $newName, ?Connection $db = null): void
+    {
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        $schema = $db->getSchema();
+        $rawOldName = $schema->getRawTableName($table);
+        $rawNewName = $schema->getRawTableName($newName);
+
+        // Rename the table
+        $db->createCommand()
+            ->renameTable($rawOldName, $rawNewName)
+            ->execute();
+
+        if ($db->getIsPgsql()) {
+            // Rename the corresponding sequence if there is one
+            // see https://www.postgresql.org/message-id/200308211224.06775.jgardner%40jonathangardner.net
+            $transaction = $db->beginTransaction();
+            try {
+                $db->createCommand()
+                    ->renameSequence("{$rawOldName}_id_seq", "{$rawNewName}_id_seq")
+                    ->execute();
+                $transaction->commit();
+            } catch (\Throwable $e) {
+                // Silently fail. The sequence probably doesn't exist
+                $transaction->rollBack();
+            }
+        }
+    }
+
+    /**
      * Returns the `id` of a row in the given table by its `uid`.
      *
      * @param string $table
@@ -1407,5 +1522,79 @@ class Db
         }
 
         return $result;
+    }
+
+    /**
+     * Looks for an index on the given table with the given columns and unique property, and returns its name,
+     * or null if no match is found.
+     *
+     * @param string $tableName The table that the index was created for.
+     * @param string|string[] $columns The column(s) that are included in the index. If there are multiple
+     * columns, separate them by commas or use an array.
+     * @param bool $unique Whether the index has a UNIQUE constraint.
+     * @param Connection|null $db The database connection.
+     * @return string|null The index name, or `null` if there isn’t a matching one.
+     * @since 4.0.0
+     */
+    public static function findIndex(string $tableName, $columns, bool $unique = false, ?Connection $db = null): ?string
+    {
+        if (is_string($columns)) {
+            $columns = StringHelper::split($columns);
+        }
+
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        foreach ($db->getSchema()->findIndexes($tableName) as $name => $index) {
+            if ($index['columns'] === $columns && $index['unique'] === $unique) {
+                return $name;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Looks for a foreign key on the given table with the given columns, and returns its name, or null if no
+     * match is found.
+     *
+     * @param string $tableName The table that the foreign key is on
+     * @param string|string[] $columns The column(s) that are included in the foreign key. If there are multiple
+     * columns, separate them by commas or use an array.
+     * @param Connection|null $db The database connection.
+     * @return string|null The foreign key name, or `null` if there isn’t a matching one.
+     * @since 4.0.0
+     */
+    public static function findForeignKey(string $tableName, $columns, ?Connection $db = null): ?string
+    {
+        if (is_string($columns)) {
+            $columns = StringHelper::split($columns);
+        }
+
+        if ($db === null) {
+            $db = self::db();
+        }
+
+        $schema = $db->getSchema();
+        $tableName = $schema->getRawTableName($tableName);
+        $table = $schema->getTableSchema($tableName);
+
+        foreach ($table->foreignKeys as $name => $fk) {
+            $fkColumns = [];
+
+            foreach ($fk as $count => $value) {
+                if ($count !== 0) {
+                    $fkColumns[] = $count;
+                }
+            }
+
+            // Could be a composite key, so make sure all required values exist!
+            if (count(array_intersect($fkColumns, $columns)) === count($columns)) {
+                return $name;
+            }
+        }
+
+        return null;
     }
 }